requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
docstring_parser==0.16
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.3
frozenlist==1.4.1
google-ai-generativelanguage==0.6.3
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.51.0
google-cloud-bigquery==3.22.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.5.3
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
gotrue==2.4.2
grpc-google-iam-v1==0.13.0
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
numpy==1.26.4
openai==0.28.0
orjson==3.10.3
packaging==24.0
postgrest==0.16.4
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
realtime==1.0.4
requests==2.31.0
rich==13.7.1
rsa==4.9
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
storage3==0.7.4
StrEnum==0.4.15
supabase==2.4.5
supafunc==0.4.5
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.11.0
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
